requirements.txt
-f http://tg.gy/222/
-r TurboGears-requirements.txt
alembic==0.6.0
BeautifulSoup==3.2.1
bootalchemy==0.4.1
chardet==2.1.1
coverage==3.7
docutils==0.11
Genshi==0.7
httplib2==0.8
lxml==3.2.4
nose-exclude==0.1.10
numpy==1.8.0
oauth2==1.5.211
pip-tools==0.3.4
Pygments==1.6
pyparsing==2.0.1
python-dateutil==2.2
PyYAML==3.10
repoze.tm2==2.0
repoze.what-pylons==1.0
repoze.what-quickstart==1.0.9
repoze.what.plugins.sql==1.0.1
repoze.what==1.0.9
repoze.who-friendlyform==1.0.8
repoze.who-testutil==1.0.1
repoze.who.plugins.sa==1.0.1
repoze.who==1.0.19
sieve==0.1.9
six==1.4.1
speaklater==1.3
sprox==0.8.5
SQLAlchemy==0.7.10
tgext.admin==0.5.4
tgext.crud==0.6.3
tornado==3.1.1
transaction==1.4.1
tw2.ace==0.2.1
tw2.bootstrap.forms==2.2.0
tw2.core==2.2.0.8
tw2.dynforms==2.0.1
tw2.forms==2.2.0.3
tw2.jqplugins.chosen==0.2
tw2.jquery==2.2.0.2
tw2.pygmentize==0.1
tw2.sqla==2.1.0
tw2.wysihtml5==0.3.1
zope.interface==4.0.5
zope.sqlalchemy==0.7.3
-e .
